+++ src/dfx/src/commands/canister/create.rs
@@ -3,7 +3,6 @@ use crate::lib::error::DfxResult;
 use crate::lib::ic_attributes::{
     get_compute_allocation, get_freezing_threshold, get_memory_allocation, CanisterSettings,
 };
-use crate::lib::identity::identity_utils::CallSender;
 use crate::lib::identity::wallet::get_or_create_wallet_canister;
 use crate::lib::operations::canister::create_canister;
 use crate::lib::root_key::fetch_root_key_if_needed;
@@ -13,6 +12,7 @@ use crate::util::clap::validators::{
 };
 use dfx_core::error::identity::IdentityError;
 use dfx_core::error::identity::IdentityError::GetIdentityPrincipalFailed;
+use dfx_core::identity::CallSender;
 
 use anyhow::{bail, Context};
 use candid::Principal as CanisterId;
