+++ src/dfx/src/commands/canister/mod.rs
@@ -1,8 +1,9 @@
 use crate::lib::agent::create_agent_environment;
 use crate::lib::error::DfxResult;
-use crate::lib::identity::identity_utils::call_sender;
 use crate::{lib::environment::Environment, NetworkOpt};
+use dfx_core::identity::CallSender;
 
+use anyhow::anyhow;
 use clap::{Parser, Subcommand};
 use tokio::runtime::Runtime;
 
@@ -64,7 +65,9 @@ pub fn exec(env: &dyn Environment, opts: CanisterOpts) -> DfxResult {
     let runtime = Runtime::new().expect("Unable to create a runtime");
 
     runtime.block_on(async {
-        let call_sender = call_sender(&opts.wallet).await?;
+        let call_sender = CallSender::determine_which(&opts.wallet)
+            .await
+            .map_err(|e| anyhow!("Failed to determine call sender: {}", e))?;
         match opts.subcmd {
             SubCommand::Call(v) => call::exec(&agent_env, v, &call_sender).await,
             SubCommand::Create(v) => create::exec(&agent_env, v, &call_sender).await,
