+++ src/dfx/src/commands/canister/uninstall_code.rs
@@ -1,8 +1,8 @@
 use crate::lib::environment::Environment;
 use crate::lib::error::DfxResult;
-use crate::lib::identity::identity_utils::CallSender;
 use crate::lib::operations::canister;
 use crate::lib::root_key::fetch_root_key_if_needed;
+use dfx_core::identity::CallSender;
 
 use candid::Principal;
 use clap::Parser;
