+++ src/dfx/src/lib/identity/mod.rs
@@ -1,2 +1,1 @@
-pub mod identity_utils;
 pub mod wallet;
